organize folder structure for better clarity and debugging
