--- app/services/flextraff.code-workspace
@@ -1,10 +1,0 @@
-{
-    "folders": [
-        {
-            "path": "../.."
-        },
-        {
-            "path": "../../../Flextraff-Admin-Panel/Panel_Frontend/flextraff_frontend"
-        }
-    ]
-}
