+++ database/002_add_rfid_logging_fields.sql
