+++ infra/docker/.dockerignore
