+++ infra/scripts/pre-push-lint.sh
