+++ tests/db/test_supabase_connection.py
