+++ tests/debug/debug_max.py
